- completed brand function, needs testing

+++ src/main/java/org/gamein/marketservergamein2022/core/sharedkernel/entity/Brand.java
@@ -24,7 +24,7 @@ public class Brand {
     private Long period;
 
     @Column(name = "brand")
-    private Integer brand;
+    private Double brand;
 
     @ManyToOne
     Team team;
